Project Finished; Ready for Review
Although no work was done over the weekend like I initially was skeptical of, I was able to finish the rest of it in today's session, including all necessary documentation (which I should have been perfectly capable of doing). The website itself even seems to work perfectly good with all the regular expressions placed in on Friday, so with that said, I believe it's safe to say this project as a whole is 100% complete (and ready for review).